+++ soc.h
@@ -173,5 +173,5 @@ uint32_t io_read(reg_type addr, reg_type* data);
 uint32_t io_write(reg_type addr, reg_type* data);
 
 uint32_t init_soc();
-uint64_t soc_tick();
+int soc_tick();
 uint64_t get_microseconds();
